Move back sirius actions to correct package
